Stuck on Settings Entity
problem: SIGABRT when either loading and/or saving SettingsEnt data

This is, I believe, because
- nil Entity coming from perisstentStore
- uninitialized peristentStore
- uninitialized SettingsEnt in persistentStore

Mostly (3) b/c AppDelegate.swift specifies only 'TrackPoint'...
maybe. But maybe that's just the name of the database.

+++ Geotify/Settings.swift
@@ -13,7 +13,7 @@ import CoreLocation
 var AppSettings:Settings = Settings()
 
 class Settings {
-  var pushAtCount:Int = 100
+  var pushAtCount:Int64 = 100
   var healthKitEnabled:Bool = true
   var batteryMonitoringEnabled:Bool = true
   var beaconMonitoringEnabled:Bool = true
